feat(usesxtokens): useSxToken -> useSxTokens
BREAKING CHANGE: useSxToken is changed to useSxTokens
