+++ src/hook/useSxTokens.test.ts
@@ -3,7 +3,7 @@ import { renderHook } from '@testing-library/react-hooks';
 import type { ThemedDict } from '../@types/ThemedDict';
 import type { ThemedTypings } from '../@types/ThemedTypings';
 
-import { useSxToken } from './useSxToken';
+import { useSxTokens } from './useSxTokens';
 
 const theme: ThemedDict = {
   colors: {
@@ -27,32 +27,32 @@ const theme: ThemedDict = {
 export function expectResult<T extends keyof ThemedTypings, V extends ThemedTypings[T]>(
   theme: ThemedDict,
   tokenGroup: T,
-  tokenValue: Exclude<V, null | undefined>,
-  expectation: any,
+  tokenValues: Array<Exclude<V, null | undefined>>,
+  expectation: any[],
 ) {
   const {
     result: { current },
-  } = renderHook(() => useSxToken(tokenGroup, tokenValue, { theme }));
+  } = renderHook(() => useSxTokens(tokenGroup, tokenValues, { theme }));
 
   return expect(current).toEqual(expectation);
 }
 
 describe('valid case', () => {
   it('colors', () => {
-    expectResult(theme, 'colors', 'red', 'red');
+    expectResult(theme, 'colors', ['red'], ['red']);
   });
 
   it('radii', () => {
-    expectResult(theme, 'radii', 'sm' as any, 8);
+    expectResult(theme, 'radii', ['sm' as any], [8]);
   });
 });
 
 describe('edge case', () => {
   it('cannot find value if tokenValue is not in tokenType', () => {
-    expectResult(theme, 'radii', '' as any, undefined);
+    expectResult(theme, 'radii', ['' as any], [undefined]);
   });
 
   it('cannot find value if theme is undefined', () => {
-    expectResult(undefined as any, 'radii', '' as any, undefined);
+    expectResult(undefined as any, 'radii', ['' as any], [undefined]);
   });
 });
